refactor(cef): carve window creation handlers into window/creation.rs
Sixth and FINAL sub-PR of the commands/window.rs modularization
(Plan 1). Pure move — no behavior change. mod.rs is now a pure
re-export shim (~46 lines).

Extracted to window/creation.rs: open_new_window, open_subwindow,
open_window_with_kind, dev_vite_port, resolve_frontend_base_url,
resolve_host_runtime_dir, assets_missing_data_url, get_offset_position,
get_secondary_window_size, and the FrontendUrlError enum + its Display/
Error impls.

mod.rs re-exports the pub setters via `pub use creation::*` and the two
pub(crate) helpers explicitly — `resolve_frontend_base_url` and
`assets_missing_data_url` are called by client / drag / window_pool /
floating_pane as `commands::window::<name>`, and a glob `pub use` only
re-exports `pub` items. Removed the now-unused top-level imports
(PathBuf / Arc / html_escape / js_string_literal / AppState) from mod.rs
— no fn bodies remain there. Also dropped an orphaned
`// ── Per-window opacity` section comment left behind by the
transparency carve.

window.rs went from 1462 LOC (one file) to a 46-line shim + six focused
sibling modules (lifecycle 376, motion 344, chrome, transparency, meta,
creation). Every external `commands::window::<name>` path is preserved.

Verified: `cargo check -p agentmux-cef --release` clean (0 errors; the
remaining warnings pre-exist on main, unrelated).
