+++ .changesets/1780033956-refactor-cef-extract-window-creation-handlers-into-commands--u2n5.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(cef): extract window creation handlers into commands/window/creation.rs (window.rs modularization complete)
